Rename DashboardDemo entry point file
